# docs/openapi-enterprise-v1.yaml
openapi: 3.0.3
info:
  title: WEPA Enterprise API
  version: 1.0.0
  description: Endpoints enterprise para documental, pagos, RBAC y auditoria.
servers:
  - url: http://localhost:8080/api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PaymentInitRequest:
      type: object
      required: [provider, obligacion_id, amount]
      properties:
        provider:
          type: string
          enum: [WOMPI, EPAYCO, MERCADOPAGO]
        obligacion_id:
          type: integer
        amount:
          type: number
        currency:
          type: string
          default: COP
        method:
          type: string
          enum: [PSE, CARD]
        payer_email:
          type: string
        payer_document:
          type: string
...
    PaymentTransferReviewRequest:
      type: object
      properties:
        observations:
          type: string
paths:
...
  /documentos/search:
    get:
      summary: Buscar documentos con filtros y paginacion server-side.
      parameters:
        - in: query
          name: entidad_tipo
          schema: { type: string }
        - in: query
          name: entidad_id
          schema: { type: integer }
        - in: query
          name: tipo_documento
          schema: { type: string }
        - in: query
          name: estado
          schema: { type: string, enum: [PENDIENTE, VALIDADO, RECHAZADO, VENCIDO] }
        - in: query
          name: nombre
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: size
          schema: { type: integer, default: 25 }
      responses:
        "200":
          description: Pagina documental.
  /documentos:
    post:
      summary: Subir documento versionado.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, entidad_tipo, entidad_id, tipo_documento]
              properties:
                file:
                  type: string
                  format: binary
                entidad_tipo:
                  type: string
                entidad_id:
                  type: integer
                tipo_documento:
                  type: string
                folder_path:
                  type: string
                fecha_vencimiento:
                  type: string
                  format: date
      responses:
        "201":
          description: Documento creado.
...
  /documentos/{id}/validar:
    patch:
      summary: Validar documento.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Documento validado.
  /documentos/{id}/estado:
    patch:
      summary: Cambiar estado documental.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Estado actualizado.
  /documentos/{id}/versiones:
    get:
      summary: Historial de versiones.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Versiones.
  /payments/initiate:
    post:
      summary: Crear transaccion de pago.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInitRequest"
      responses:
        "201":
          description: Link/estado de pago.
  /payments/obligacion/{id}:
    get:
      summary: Listar transacciones por obligacion.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Transacciones.
  /payments/transfers:
    get:
      summary: Listar pagos por transferencia para revision administrativa.
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [PENDING_VERIFICATION, APPROVED, REJECTED]
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: size
          schema: { type: integer, default: 25 }
      responses:
        "200":
          description: Transferencias filtradas.
  /payments/transfers/{id}:
    get:
      summary: Detalle de transferencia.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Transferencia.
  /payments/transfers/{id}/history:
    get:
      summary: Historial de validaciones de transferencia.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Eventos de validacion.
  /payments/transfers/{id}/support:
    get:
      summary: Soporte autenticado de transferencia.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Archivo de soporte.
  /payments/transfers/{id}/receipt:
    get:
      summary: Recibo PDF de transferencia aprobada.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Recibo PDF.
  /payments/transfers/{id}/approve:
    patch:
      summary: Aprobar transferencia, descontar saldo y generar recibo.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentTransferReviewRequest"
      responses:
        "200":
          description: Transferencia aprobada.
  /payments/transfers/{id}/reject:
    patch:
      summary: Rechazar transferencia con observaciones.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentTransferReviewRequest"
      responses:
        "200":
          description: Transferencia rechazada.
...